Apply theme to app pages
