+++ src/components/Apps/AppInfoSnippet.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   AppCircle,
   AppCircleContainer,
@@ -11,7 +10,7 @@ import {
   AppInfoSnippetRight,
   AppInfoUserName,
 } from './Apps-styles';
-import { Avatar, ButtonBase } from '@mui/material';
+import { Avatar, ButtonBase, useTheme } from '@mui/material';
 import { getBaseApiReact } from '../../App';
 import LogoSelected from '../../assets/svgs/LogoSelected.svg';
 import { Spacer } from '../../common/Spacer';
@@ -41,6 +40,8 @@ export const AppInfoSnippet = ({
   const setSettingsLocalLastUpdated = useSetRecoilState(
     settingsLocalLastUpdatedAtom
   );
+  const theme = useTheme();
+
   return (
     <AppInfoSnippetContainer
       sx={{
@@ -96,6 +97,7 @@ export const AppInfoSnippet = ({
             </AppCircle>
           </AppCircleContainer>
         </ButtonBase>
+
         <AppInfoSnippetMiddle>
           <ButtonBase
             onClick={() => {
@@ -112,12 +114,17 @@ export const AppInfoSnippet = ({
           >
             <AppInfoAppName>{app?.metadata?.title || app?.name}</AppInfoAppName>
           </ButtonBase>
+
           <Spacer height="6px" />
+
           <AppInfoUserName>{app?.name}</AppInfoUserName>
+
           <Spacer height="3px" />
+
           <AppRating app={app} myName={myName} />
         </AppInfoSnippetMiddle>
       </AppInfoSnippetLeft>
+
       <AppInfoSnippetRight
         sx={{
           gap: '10px',
@@ -157,7 +164,7 @@ export const AppInfoSnippet = ({
             setSettingsLocalLastUpdated(Date.now());
           }}
           sx={{
-            backgroundColor: '#359ff7ff',
+            backgroundColor: theme.palette.background.paper,
             opacity: isSelectedAppPinned ? 0.6 : 1,
           }}
         >
@@ -174,7 +181,9 @@ export const AppInfoSnippet = ({
             });
           }}
           sx={{
-            backgroundColor: isInstalled ? '#0091E1' : '#247C0E',
+            backgroundColor: isInstalled
+              ? '#0091E1'
+              : theme.palette.background.paper,
           }}
         >
           <AppDownloadButtonText>
